# semantic-tokens: cache modifier lookups