Make a buffer-local cache of the mapping from a modifier bitmap to the
set of faces it resolves to.

The assumption is that for a given language the modifier usage will
tend to cluster, so the lookup table will not be large.