@sebastiansturm 2022-08-15
Hi @alanz, thanks for your PR!
I tried it out today, but so far haven't seen performance improvements; on the contrary, performance seemed to get worse in the one scenario I tested. To be precise, I cloned the latest rust-analyzer from source and, within crates/hir-def/src/nameres/collector.rs, executed the following function (using a very recent Emacs 29 build, with lsp-mode natively compiled in both cases):
```
(defun benchmark/fontification (count) ()
       (garbage-collect)
       (benchmark-call (lambda () (lsp-semantic-tokens--fontify #'font-lock-fontify-region (point-min) (point-max))) count))
```
The results of 5 consecutive evaluations of `(benchmark/fontification 50)` were as follows:
```
current lsp-mode master:
(4.812272216 3 0.34256550100000016)
(4.8488806 3 0.3737376499999998)
(4.849987764000001 3 0.3759672280000004)
(4.850168844 3 0.3760423690000003)
(4.843063498999999 3 0.3765382319999997)

PR 3670:
(5.1777232269999995 6 0.536172971)
(5.1809434 6 0.5332634989999998)
(5.182143025 6 0.5344310930000002)
(5.177091014 6 0.5340501979999996)
(5.180054484 6 0.5352050740000003)
```
One obvious pessimization (that you already remarked on in a comment, but AFAICS didn't address yet) is that empty modifier face sets will fail the `unless faces-to-apply` check and thus get recomputed every time. Perhaps fixing that will lead to actual performance improvements? Or maybe you've been testing a different use case that already benefits from the PR in its current form; if that is the case, what benchmarks should I perform to reproduce your results?