@alanz 2022-08-15
Thanks for taking a look @sebastiansturm .

I am actually using it from my https://github.com/alanz/lsp-mode/tree/semantic-modifiers-optimise branch, which is basically this PR on top of https://github.com/emacs-lsp/lsp-mode/pull/3668, which fully populates the modifier faces for rust-analyzer.

And I put the optimisation in pace because otherwise navigating in rust files such as rust-analyzer `main_loop.rs` was terribly slow.  I did not benchmark it, but found a noticeable sluggishness doing something simple like going to the top of that file (once the server was fully started up), and just moving down line by line with my cursor.  It is not nearly so sluggish using this change.