@alanz 2022-08-15
I just did those benchmarks on my own machine with emacs 28.1, jit enabled, without precompiling lsp-rust.el or lsp-semantic-tokens.el.

I get

```
On my semantic-modifiers-optimise branch
With bad font specifications
(0.003357608 0 0.0)
(0.003357608 0 0.0)
(0.003357608 0 0.0)

Without the semantic token optimisation
With bad font specifications
(18.839329314 32 1.6363554149999997)
(18.839329314 32 1.6363554149999997)

Same, but with good font specifications
(18.737020591 32 1.626868912)
(18.620734945 32 1.61827597)
(18.620734945 32 1.61827597)
```

My *Messages* was littered with missing font declarations, so I changed lsp-rust.el to have 

```elisp
(defun lsp-rust-analyzer--set-tokens ()
  "Set the mapping between rust-analyzer keywords and fonts to apply.
The keywords are sent in the initialize response, in the semantic
tokens legend."
  (setq lsp-semantic-token-modifier-faces
        '(
          ("documentation" . font-lock-nop-face)
          ("declaration" . font-lock-nop-face)
          ("definition" . font-lock-nop-face)
          ("static" . font-lock-nop-face)
          ("abstract" . font-lock-nop-face)
          ("deprecated" . font-lock-nop-face)
          ("readonly" . font-lock-nop-face)
          ("default_library" . font-lock-nop-face)
          ("async" . font-lock-nop-face)
          ("attribute" . font-lock-nop-face)
          ("callable" . font-lock-nop-face)
          ("constant" . font-lock-nop-face)
          ("consuming" . font-lock-nop-face)
          ("control_flow" . font-lock-nop-face)
          ("crate_root" . font-lock-nop-face)
          ("injected" . font-lock-nop-face)
          ("intra_doc_link" . font-lock-nop-face)
          ("library" . font-lock-nop-face)
          ("mutable" . font-lock-mutable-modifier-face)
          ("public" . font-lock-nop-face)
          ("reference" . font-lock-reference-modifier-face)
          ("trait" . font-lock-nop-face)
          ("unsafe" . font-lock-nop-face)
          )))
```
Which shut them up, but made no difference to the performance.

So for me it makes a *big* difference to use this PR