@alanz 2022-08-16
I revisited this tonight, and redid my benchmarks.  And realised my first result in https://github.com/emacs-lsp/lsp-mode/pull/3670#issuecomment-1215811793 is bogus.

Herewith revised benchmark results

```
Benchmark, on collector.rs
(benchmark/fontification 50)

My updated PR 3668 (lsp-rust semantic modifiers, this PR not applied)
(18.640859705 28 1.5243770410000002)
(18.620477035 28 1.5261523759999998)
(18.438429331 28 1.5186607939999996)

My caching, on top of PR 3668. State of this PR as of yesterday
(14.82362226 54 2.765928829)
(14.771710461000001 53 2.7048730019999994)
(14.812715127 53 2.7507432979999997)

My improved caching, on top of PR 3668, state as of now
(8.524490870000001 43 2.191504358)
(8.454953522 42 2.112336408)
(8.417772501 42 2.1134970299999996)
```
The change I made was to supply a symbol for the not-found state in the hashmap lookup to distinguish between not cached and no fonts applicable.

And with that it goes roughly twice as fast as without it.