@sebastiansturm 2022-08-16
thanks! Makes sense that caching would help with that long list of modifiers, and might be a good idea in general, now that the no-modifiers case is no longer being pessimized. Adding a few inline comments now...